dev
```
               Method |          Mean |     StdDev | Scaled |          RPS | Allocated |
--------------------- |-------------- |----------- |------- |------------- |---------- |
 PlaintextTechEmpower |   531.1228 ns |  5.7545 ns |   1.00 | 1,882,803.68 |     112 B |
           LiveAspNet |   917.0677 ns |  9.0826 ns |   1.73 | 1,090,432.07 |     112 B |
              Unicode | 1,300.7374 ns | 18.2239 ns |   2.45 |   768,794.70 |     112 B |
```
pr
```
               Method |          Mean |     StdDev | Scaled |          RPS | Allocated |
--------------------- |-------------- |----------- |------- |------------- |---------- |
 PlaintextTechEmpower |   513.2275 ns |  6.5310 ns |   1.00 | 1,948,453.72 |     112 B |
           LiveAspNet |   909.4809 ns | 10.0160 ns |   1.77 | 1,099,528.33 |     112 B |
              Unicode | 1,283.9817 ns | 13.0742 ns |   2.50 |   778,827.32 |     112 B |
```