@benaadams 2017-03-08
pr
```
               Method |          Mean |     StdDev | Scaled |          RPS | Allocated |
--------------------- |-------------- |----------- |------- |------------- |---------- |
 PlaintextTechEmpower |   503.9617 ns |  6.1930 ns |   1.00 | 1,984,277.87 |     112 B |
           LiveAspNet |   901.9177 ns |  9.6750 ns |   1.79 | 1,108,748.57 |     112 B |
              Unicode | 1,258.5205 ns | 14.2006 ns |   2.50 |   794,583.77 |     112 B |
```