@benaadams 2017-03-09
dev
```
                        Method |      Mean |    StdDev | Scaled |        RPS | Allocated |
------------------------------ |---------- |---------- |------- |----------- |---------- |
          PlaintextTechEmpower | 1.4231 us | 0.0063 us |   1.00 | 702,705.85 |     353 B |
 PipelinedPlaintextTechEmpower | 1.1439 us | 0.0076 us |   0.80 | 874,220.46 |     351 B |
                    LiveAspNet | 3.0643 us | 0.0362 us |   2.15 | 326,334.31 |   1.12 kB |
           PipelinedLiveAspNet | 2.7153 us | 0.0256 us |   1.91 | 368,284.39 |   1.11 kB |
                       Unicode | 5.2271 us | 0.0370 us |   3.67 | 191,309.20 |   1.99 kB |
              UnicodePipelined | 5.0083 us | 0.1999 us |   3.52 | 199,669.47 |   1.98 kB |
```
pr
```
                        Method |      Mean |    StdDev | Scaled |        RPS | Allocated |
------------------------------ |---------- |---------- |------- |----------- |---------- |
          PlaintextTechEmpower | 1.3217 us | 0.0070 us |   1.00 | 756,626.99 |     349 B |
 PipelinedPlaintextTechEmpower | 1.0927 us | 0.0088 us |   0.83 | 915,173.46 |     348 B |
                    LiveAspNet | 2.8567 us | 0.0300 us |   2.16 | 350,049.41 |   1.11 kB |
           PipelinedLiveAspNet | 2.5057 us | 0.0140 us |   1.90 | 399,089.12 |    1.1 kB |
                       Unicode | 4.8917 us | 0.0261 us |   3.70 | 204,427.38 |   1.95 kB |
              UnicodePipelined | 4.7156 us | 0.0321 us |   3.57 | 212,062.52 |   1.97 kB |
```